Significantly lower memory usage
There was a significant amount of memory being used to simply
represent the filenames in memory.  By using an integer and
keeping a mapping between it and the name we can reduce the
amount of memory to a more reasonable amount of consumption.

Running against the Linux kernel source tree took 17.4GiB before
this change.  With this change we max out at ~4.8GiB.

Signed-off-by: Tony Asleson <tony.asleson@gmail.com>