Ouput json when requested
This gets us JSON output, but it's not easy to consume.  The
memory saving stuff that is in place where we represent file names
by an integer ends up getting exposed here.  Thus the consumer
needs to use "index_to_name" to figure out which file is which.  I
think I need to re-work the serialization process to make this
easier for others to consume.  The main issue is I don't want
to stuff a reference to the FileId in every single struct
Collision so that I can convert from index to file name when
I serialize a Collision.  Need a way to make this global data
global safely.

Ref. https://github.com/tasleson/duplihere/issues/4

Signed-off-by: Tony Asleson <tony.asleson@gmail.com>