Do rolling hash insert in thread
When we a processing each of the files in threads lets go ahead and
insert the file signatures into the file_hashes vector and add all
the rolling hash values into the shared file hashes struct.  This
is likely creating a lot of lock contention on the collisions hash.

Performance quite a bit better, but need to understand why the CPU
utilization is being capped to 2 cores early in the process.  Could
be the lock contention causing things to get serialized between a
producer and a consumer.

Signed-off-by: Tony Asleson <tony.asleson@gmail.com>