Add some more comments
I should have taken better notes during development as I'm already
starting to question the why we are doing things a certain way. At
the time it seemed very clear, months later, not so much.

Using integers to represent files was a huge improvement in
memory reduction and performance by removing string compares, but
it sure made the code harder to follow.

Investigate how we can make this clearer.

Signed-off-by: Tony Asleson <tony.asleson@gmail.com>